# feat: Session management performance optimizations for high-throughput environments