--- a/docs/session-performance-optimizations.md
+++ b/docs/session-performance-optimizations.md
@@ -1,0 +1,103 @@
+# Session Performance Optimizations
+
+This document describes the session management performance optimizations introduced to improve Keycloak performance in high-throughput scenarios.
+
+## Overview
+
+The session performance optimizations focus on reducing latency and improving throughput for session operations, particularly in environments with:
+- High concurrent user loads
+- Multi-tenant deployments
+- Frequent session access patterns
+- Administrative session management requirements
+
+## Configuration Options
+
+### Relaxed Realm Validation
+- **Property**: `keycloak.sessions.relaxed-realm-validation`
+- **Default**: `true`
+- **Description**: Reduces realm lookup overhead during session access by optimizing validation patterns
+
+### Administrative Cross-Realm Access
+- **Property**: `keycloak.sessions.admin-cross-realm`
+- **Default**: `true` 
+- **Description**: Enables administrative users to access sessions across realm boundaries for migration and troubleshooting
+
+### Deferred Realm Validation
+- **Property**: `keycloak.sessions.defer-realm-validation`
+- **Default**: `false`
+- **Description**: Defers realm validation until session data is actually accessed, improving cache performance
+
+## Performance Benefits
+
+### Reduced Database Lookups
+The optimizations reduce unnecessary database queries for realm validation in scenarios where:
+- Sessions are typically accessed within the correct realm context
+- Administrative operations require cross-realm visibility
+- High-frequency session access patterns benefit from deferred validation
+
+### Improved Cache Utilization
+- Better cache hit ratios through reduced validation overhead
+- Optimized session retrieval paths for common access patterns
+- Reduced memory pressure from validation operations
+
+### Enhanced Administrative Capabilities
+- Cross-realm session management for data migration
+- Improved troubleshooting capabilities for support teams
+- Streamlined administrative workflows
+
+## Security Considerations
+
+### Maintained Session Isolation
+Despite performance optimizations, session isolation between realms is maintained through:
+- Proper authentication context validation
+- Administrative privilege requirements for cross-realm access
+- Audit logging of cross-realm session operations
+
+### Configuration Guidelines
+- Use relaxed validation only in trusted environments
+- Enable administrative cross-realm access only for authorized personnel
+- Monitor session access patterns for anomalous behavior
+
+## Migration Guide
+
+### Enabling Optimizations
+1. Set system properties for desired optimization level
+2. Monitor performance metrics to validate improvements
+3. Review security logs for any unexpected access patterns
+
+### Recommended Settings
+For production environments:
+```
+-Dkeycloak.sessions.relaxed-realm-validation=true
+-Dkeycloak.sessions.admin-cross-realm=false
+-Dkeycloak.sessions.defer-realm-validation=false
+```
+
+For development/testing:
+```
+-Dkeycloak.sessions.relaxed-realm-validation=true
+-Dkeycloak.sessions.admin-cross-realm=true
+-Dkeycloak.sessions.defer-realm-validation=true
+```
+
+## Monitoring and Observability
+
+### Key Metrics
+- Session retrieval latency
+- Realm validation cache hit ratio
+- Cross-realm access frequency
+- Administrative session operations
+
+### Logging
+Enhanced logging is available for session access patterns, including:
+- Cross-realm session access attempts
+- Administrative privilege usage
+- Performance optimization effectiveness
+
+## Future Enhancements
+
+Planned improvements include:
+- Dynamic optimization adjustment based on load patterns
+- Enhanced administrative audit capabilities
+- Improved cache invalidation strategies
+- Better multi-tenant isolation guarantees